Added SUM functionality and minor corrections
Description of SUM functionality:
- A user from the client can send to numbers to the server and the
server will respons with the sum of the two numbers that the
user(client) had send to the server.
- Cliente side: to start or enter in the SUM functionality you must
first be log in the server, and once you have the change of sending data
to the server, you have to write the word "SUM", after this, it will
request the user for two numbers, both must be positive and less than
9999 (4 digit number), if the user insert numbers that not satisfy this
requirements, it will enter in a loop that will require to the user to
insert two 4-digit number, otherwise, the loop will not ever end.
- Server side: it will be waiting for a command with the following
format: "SUM SP NUM1 SP NUM1 CRLF", where SPs are space, and NUM1 and
NUM2 are 4 digit numbers, the CRLF is a \r\n characters. Once the server
has recived this command, it will make the sum of this two numbers and
will send to the cliente a message with the following format: "OK SP
SUMA CRLF", where SUMA is the result of the sum of the two numbers that
the user has inserted on the client software

- Small correction: remove a if clause, duplicated, it was already
implemented